fgadsjnkrfddfgsvsdser fsfd

+++ src/ronin-world/Battlegrounds/BattlegroundMgr.h
@@ -310,7 +310,7 @@ public:
     void SendPVPData(Player* plr);
 
     /* Get the starting position for this team. */
-    virtual LocationVector GetStartingCoords(uint32 Team) = 0;
+    virtual LocationVector GetStartingCoords(uint8 Team) = 0;
 
     /* Send a packet to the entire battleground */
     void DistributePacketToAll(WorldPacket * packet);
