+++ src/ronin-world/Battlegrounds/Expansion1/EyeOfTheStorm.h
@@ -27,7 +27,7 @@ public:
     void HookOnHK(Player* plr);
     void HookOnShadowSight();
     void SpawnBuff(uint32 x);
-    LocationVector GetStartingCoords(uint32 Team);
+    LocationVector GetStartingCoords(uint8 Team);
     static CBattleground* Create( MapMgr* m, uint32 i, uint32 l, uint32 t) { return new EyeOfTheStorm(m, i, l, t); }
     WoWGuid GetFlagHolderGUID() { return m_flagHolder; }
